requirements.txt
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Jinja2==3.1.2
PyYAML==5.4.1
Flask-MySQLdb==1.0.1
